refactor(agent): extract OpenAI usage/pricing into pure functions
Replicate the Anthropic usage/pricing extraction for the OpenAI handler
(base for the 6 OpenAI-compatible providers). New openai/openAIUsage.ts:

- computeOpenAIPrice     (prompt/completion + reasoning + cache-discount)
- normalizeOpenAIUsage   (unified NormalizedUsage shape)

Pure functions of the raw usage plus an OpenAIPricingConfig
(inputPrice/outputPrice/cacheDiscountFactor); normalizeOpenAIUsage also
takes the provider id so the usageProvider override (e.g. Kimi) still flows
through. The handler keeps computePrice/normalizeUsage as thin delegators —
the IModelHandler surface is unchanged, and all 6 compatible subclasses
inherit the delegators unchanged.

No behavior change. Verified: typecheck, lint (0 errors), full test:kernel
(1383 tests) pass.
